PIO now build in serial.  Still testing, but the serial build does appear to work properly so far.
The problem is still a mystery to me.  With the serial build,
CMake was getting the dependency order wrong, and not building
the timing library before building pio.  I added an 'add_dependencies'
line to the pio/CMakeLists.txt file to force CMake to get the
dependency right.

Most of the rest of the changes are minor (formatting) changes.
I did add an ADDITIONAL_INCS variable to pass in additional
include directories, since just passing in the mpi-serial
location via CMAKE_Fortran_FLAGS does not suffice for the
C-build of GPTL.  Just seemed like the right thing to do.

I added debug flags to the config.pl script and added the
sermake.sh script to build in serial.  Some other script
changes just to work with my directory structures, but they
are very minor.